🔨 move fetch hooks to separate file
