+++ packages/@ourworldindata/utils/src/index.ts
@@ -40,6 +40,7 @@ export {
     urlToSlug,
     trimObject,
     fetchText,
+    fetchJson,
     getUserCountryInformation,
     stripHTML,
     getRandomNumberGenerator,
